MSBuild: Add 64-bit builds for the dev & client tools
The tools mostly had builds "added" previously, but now we've gone over and ensured behaviour is correct in 64-bit builds.

+++ Client/KscViewer/KscViewerDoc.cpp
@@ -94,7 +94,7 @@ BOOL CKscViewerDoc::OnOpenDocument(LPCTSTR lpszPathName)
 		return FALSE;
 
 	char* szExt = nullptr;
-	int nLen = strlen(lpszPathName);
+	size_t nLen = strlen(lpszPathName);
 
 	szExt = (char*)lpszPathName + nLen - 3;
